Rename store_data -> store
