+++ src/kuadrant/pipeline/tasks/store.rs
